Move SD Card Menu to its own file
